Rewrote ContainerState, improved network servers
ContainerStateServer is a system to allow the current TASing state to sync between clients on server join.

But since you can start e.g. recording in the main menu, and the server is not started or there is no connection to the server, we can't send the correct state to the server yet. Therefore we send the state when the client joins the server.

Previously, the server would "request" the state from the client, now the client will send the packet regardless and the server will interpret if the state should be changed or not.

- Switched to the new packet server for container states
- Improved sending packets while the custom server is starting. Now the client will wait, until the server send a packet back, to ensure it being started
- Implemented ContainerStateServer.leaveServer() which I appearently forgot